Make verified mean verified
By that, I mean that this package now tries to identify both incorrect
and correct uses of whatever is described in the rules given to it. It
doesn't just blindly "verify" every single phrase matched by a simple
pattern.

The intention is that one should first use proofreadWith to highlight
probable mistakes and non-mistakes, and then optionally highlightWith to
highlight special characters that cannot be visually distinguished from
their incorrect counterparts.